🎨 adjusted the structure of lib
